Update to make all enumeration classes.
Implemented our own time parser so we properly support time widgets.
Fixed the plugin_exception so it derives from the snap_exception.
Fixed some controlled variables that limited enumerations using auto_init since they had not yet been updated to auto_enum_init.
Removed some static casts that were necessary to copy enumerations.
Added some static casts in cases where we need to extract the value of an enumeration.
Fixed formatting in all headers, as much as I could do in one go.
Changed the word_t enumeration to have upper case.
Replaced a couple of #define into static variables.

+++ snapwebsites/lib/table_model.cpp
@@ -91,7 +91,7 @@ void table_model::fetchMore(QModelIndex const & model_index)
         //f_table->clearCache();
         f_rowsRemaining = f_table->readRows( f_rowp );
 
-        int const itemsToFetch( qMin(f_rowCount, f_rowsRemaining) );
+        int32_t const itemsToFetch( qMin(static_cast<int32_t>(f_rowCount), static_cast<int32_t>(f_rowsRemaining)) );
 
         beginInsertRows( QModelIndex(), f_pos, f_pos + itemsToFetch - 1 );
         endInsertRows();
